Major refactoring of classes, added decorator module
Refactored classes in classes.py to SkyComponent and SkyModel class (dealing with combination of multiple components). These inherit from the superclass _BaseSkyClass which is an abstract class (t_b being the only abstract method). All child classes of the SkyComponent class inherit from it. Decorator module also added with decorators allowing modification of method docstrings with current values of class constants etc. Future decorators are to be put in this module.